Correct networkd enablement check to use booleans
`when` wants a boolean and using the dicts does evaluate to `True` in a boolean
context. Newer versions of ansible however have decided that it wants us to give
it an actual boolean instead of something that can become one.

Check explicitly for the dicts being defined by adding `is defined` to each of
the named dicts.